Move extract resources to its pkg
Move ExtractContainerResourceValue
