+++ pkg/fieldpath/BUILD
@@ -15,12 +15,7 @@ go_library(
         "fieldpath.go",
     ],
     tags = ["automanaged"],
-    deps = [
-        "//pkg/api:go_default_library",
-        "//pkg/api/v1:go_default_library",
-        "//vendor:k8s.io/apimachinery/pkg/api/meta",
-        "//vendor:k8s.io/apimachinery/pkg/api/resource",
-    ],
+    deps = ["//vendor:k8s.io/apimachinery/pkg/api/meta"],
 )
 
 go_test(
@@ -30,8 +25,6 @@ go_test(
     tags = ["automanaged"],
     deps = [
         "//pkg/api/v1:go_default_library",
-        "//vendor:github.com/stretchr/testify/assert",
-        "//vendor:k8s.io/apimachinery/pkg/api/resource",
         "//vendor:k8s.io/apimachinery/pkg/apis/meta/v1",
     ],
 )
