+++ pkg/fieldpath/fieldpath_test.go
@@ -20,9 +20,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
-
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/kubernetes/pkg/api/v1"
 )
@@ -116,122 +113,6 @@ func TestExtractFieldPathAsString(t *testing.T) {
 			}
 		} else if e := tc.expectedValue; e != "" && e != actual {
 			t.Errorf("%v: unexpected result; got %q, expected %q", tc.name, actual, e)
-		}
-	}
-}
-
-func getPod(cname, cpuRequest, cpuLimit, memoryRequest, memoryLimit string) *v1.Pod {
-	resources := v1.ResourceRequirements{
-		Limits:   make(v1.ResourceList),
-		Requests: make(v1.ResourceList),
-	}
-	if cpuLimit != "" {
-		resources.Limits[v1.ResourceCPU] = resource.MustParse(cpuLimit)
-	}
-	if memoryLimit != "" {
-		resources.Limits[v1.ResourceMemory] = resource.MustParse(memoryLimit)
-	}
-	if cpuRequest != "" {
-		resources.Requests[v1.ResourceCPU] = resource.MustParse(cpuRequest)
-	}
-	if memoryRequest != "" {
-		resources.Requests[v1.ResourceMemory] = resource.MustParse(memoryRequest)
-	}
-	return &v1.Pod{
-		Spec: v1.PodSpec{
-			Containers: []v1.Container{
-				{
-					Name:      cname,
-					Resources: resources,
-				},
-			},
-		},
-	}
-}
-
-func TestExtractResourceValue(t *testing.T) {
-	cases := []struct {
-		fs            *v1.ResourceFieldSelector
-		pod           *v1.Pod
-		cName         string
-		expectedValue string
-		expectedError error
-	}{
-		{
-			fs: &v1.ResourceFieldSelector{
-				Resource: "limits.cpu",
-			},
-			cName:         "foo",
-			pod:           getPod("foo", "", "9", "", ""),
-			expectedValue: "9",
-		},
-		{
-			fs: &v1.ResourceFieldSelector{
-				Resource: "requests.cpu",
-			},
-			cName:         "foo",
-			pod:           getPod("foo", "", "", "", ""),
-			expectedValue: "0",
-		},
-		{
-			fs: &v1.ResourceFieldSelector{
-				Resource: "requests.cpu",
-			},
-			cName:         "foo",
-			pod:           getPod("foo", "8", "", "", ""),
-			expectedValue: "8",
-		},
-		{
-			fs: &v1.ResourceFieldSelector{
-				Resource: "requests.cpu",
-			},
-			cName:         "foo",
-			pod:           getPod("foo", "100m", "", "", ""),
-			expectedValue: "1",
-		},
-		{
-			fs: &v1.ResourceFieldSelector{
-				Resource: "requests.cpu",
-				Divisor:  resource.MustParse("100m"),
-			},
-			cName:         "foo",
-			pod:           getPod("foo", "1200m", "", "", ""),
-			expectedValue: "12",
-		},
-		{
-			fs: &v1.ResourceFieldSelector{
-				Resource: "requests.memory",
-			},
-			cName:         "foo",
-			pod:           getPod("foo", "", "", "100Mi", ""),
-			expectedValue: "104857600",
-		},
-		{
-			fs: &v1.ResourceFieldSelector{
-				Resource: "requests.memory",
-				Divisor:  resource.MustParse("1Mi"),
-			},
-			cName:         "foo",
-			pod:           getPod("foo", "", "", "100Mi", "1Gi"),
-			expectedValue: "100",
-		},
-		{
-			fs: &v1.ResourceFieldSelector{
-				Resource: "limits.memory",
-			},
-			cName:         "foo",
-			pod:           getPod("foo", "", "", "10Mi", "100Mi"),
-			expectedValue: "104857600",
-		},
-	}
-	as := assert.New(t)
-	for idx, tc := range cases {
-		actual, err := ExtractResourceValueByContainerName(tc.fs, tc.pod, tc.cName)
-		if tc.expectedError != nil {
-			as.Equal(tc.expectedError, err, "expected test case [%d] to fail with error %v; got %v", idx, tc.expectedError, err)
-		} else {
-			as.Nil(err, "expected test case [%d] to not return an error; got %v", idx, err)
-			as.Equal(tc.expectedValue, actual, "expected test case [%d] to return %q; got %q instead", idx, tc.expectedValue, actual)
 		}
 	}
 }
